// ui/material-ui/src/app/core/enums/gameModeType.enum.ts
export enum GameModeType {
    FIVE_BASIC = "FIVE_BASIC",
    FIVE_MORDRED_PERCIVAL = "FIVE_MORDRED_PERCIVAL",
    FIVE_MORGANA_PERCIVAL = "FIVE_MORGANA_PERCIVAL",
    SIX_BASIC = "SIX_BASIC",
    SEVEN_BASIC = "SEVEN_BASIC",
    EIGHT_BASIC = "EIGHT_BASIC",
    NINE_BASIC = "NINE_BASIC",
    TEN_BASIC = "TEN_BASIC"
}